pyproject.toml
[tool.poetry]
name = "ivision"
version = "1.0.0"
description = "A library for utilizing Apple's Vision Framework from Python in an easy manner"
authors = ["Tim Wedde <timwedde@icloud.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/timwedde/ivision"
homepage = "https://github.com/timwedde/ivision"

[tool.poetry.scripts]
ivision = "ivision.ivision:cli"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
pyobjc-framework-Vision = "^7.1"
pandas = "^1.2.0"
rich = "^9.5.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest = "^6.0.1"
black = "^20.8b1"
pre-commit = "^2.9.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
